add read info only from cache!

+++ rosie/src/main/java/com/karumi/rosie/repository/PaginatedRosieRepository.java
@@ -86,13 +86,15 @@ public class PaginatedRosieRepository<K, V extends Identifiable<K>> extends Rosi
     for (PaginatedCacheDataSource<K, V> cacheDataSource : paginatedCacheDataSources) {
       values = cacheDataSource.getPage(page);
 
-      if (values != null) {
+      if (values != null && values.getItems() != null && values.getItems().size() > 0) {
         if (areValidValues(values, cacheDataSource)) {
           break;
         } else {
           cacheDataSource.deleteAll();
           values = null;
         }
+      } else {
+        values = null;
       }
     }
